backend/requirements.txt
asgiref==3.7.2 
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.3
django-cors-headers==4.1.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-base64==2.0
drf-writable-nested==0.7.0
idna==3.4
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.0
oauthlib==3.2.2
pandas==2.0.3
Pillow==10.0.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2023.3
reportlab==4.0.4
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
SQLAlchemy==2.0.19
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
webcolors==1.11.1
gunicorn==21.2.0
